+++ semver-settings-gradle-plugin/testFunctional/kotlin/com/javiersc/semver/settings/gradle/plugin/utils/createGitIgnore.kt
@@ -1,0 +1,21 @@
+package com.javiersc.semver.settings.gradle.plugin.utils
+
+import java.io.File
+
+fun File.createGitIgnore() {
+    resolve(".gitignore").apply {
+        createNewFile()
+        writeText(
+            """
+                |.gradle/
+                |.idea/
+                |.kotlin/
+                |build/
+                |local.properties
+                |environment/
+                |
+            """
+                .trimMargin()
+        )
+    }
+}